requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
